mount only files and directories listed in /etc/files
Previously, we scanned /config for all files and bind mounted
the files directly.  This lead to some inconsistencies between
reboots.  A file that is persisted because it's parent directory
is persisted in bind mounted directly after a reboot rather than
it's parent directory.  This causes problems with tools like
unlink or rename that can't access the file when it's bind
mounted.

Also, a directory that is persisted with no files inside would
end up not being created/bind-mounted on reboot

rhbz#860279

Change-Id: I6dce8d619b8a6a51d636974eb34141f8aac9bae7
Signed-off-by: Mike Burns <mburns@redhat.com>