fix list assignment in rhn code
in python, when variable "a" is a list, and you assign

b=a

you get 2 variables pointing to the same list.  So any
operation you do on a, get done on b and vice versa.
Instead of this, do b=list(a) to get a new list with the
same content.

Change-Id: I3ac8365d605dd664b74243bd76889b1c9e41e608
Signed-off-by: Mike Burns <mburns@redhat.com>