dracut: Expand and translate devices to compare
The storage_init parameter is a set of devices which pontentially
contain PVs which in turn are members of VGs. Before these VGs are
removed it needs to be determined if all PVs part of the given VGs are
covered by the devices in the storage_init karg.
In order to do this some comparisons are needed, but these comparisons
need to take place after (a) the storage_init values are parsed and (b)
the device names are translated into the device names which are used by
LVM.

Previously the parsing and translation did not happen in some
comparisons which broke the logic.
Two passes are introduced which first parse and afterwards translate the
device names to allow a comparison.

rhbz#869172

Change-Id: I8eb9cff20ef9a0330451c455f35cfbbccdf27a96
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>