storage: Check for remaining HostVG
Previously it was assumed that no HostVG existed after all the wiping
was done, but this could be the case when the PV was on a disk which was
not wiped.
Now a message is added to the logs which points at the cause for this
problem and the installation is aborted.

rhbz#872114

Change-Id: Ia9eea77f867adfd601d1dd1d3fcf46d6b07e5414
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>