storage: Translate disk to name used by LVM
Even if e.g. /dev/sd* is passed to LVM, LVM might decide to use one of
the other dm paths pointing to that device. This raises problems when
the original name of the disk is expected to turn up in the output of
LVM, but this assumption is wrong, as LVM silently translates names.

rhbz#872502

Change-Id: I8a628fcc8c213172bc13a849a5ea9513381faa8b
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>