functions: Improve virt capability detection
Previously it was just searched for the term "kvm" in the capabilities
provided by libvirt to determin if kvm is enabled or not. This could
lead to false positives as the term "kvm" did not only appear as a
domain type (the information the logic is looking for) but also as part
of the emulator (qemu-kvm).
Now the search is more precise and kvm support is only detected if a kvm
domain is found.

rhbz#874377

Change-Id: I0f45a14c3de46c52f2f04b54af049872fb39280d
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>