rhn: Make rhn_check file based
Previously the rhn_check scripts was used to determin if a Node was
connected to RHN, this could lead to slow page loading times in cases
where RHN or the network connection was slow. The new approach looks for
a specififc file which exists if the Node is registered.

rhbz#816900

Change-Id: I9042f99628bded3adb5b9827202762e399dfa28f
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>