installer: Fix other-disk disk selection
Previously the installer tried to populate the disk details widget on
the root device page even when "Other Device" was selected. This failed
and raised an exception.
With this patch the installer only shows details when the device is show
in the list and not the "other device" entry.

rhbz#895413

Change-Id: I072638aea4ed2394b9c0e1529a04743a94a9250d
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>