setup: Discard kdump message after displaying it
Previously every error message which appeared on the kdump page wasn't
discarded after it was shown, this way the error message was shown even
if it did not apply. E.g. when kdump was configured with NFS and th
epath was missing an appropriate message was shown, if in a second try
the restore option was used the NFS related message was still shown.
This patch discards kdump error messages after they were shown.

rhbz#903989

Change-Id: I383d3ac0b0f6b68633a2a7c67a2dc380989e472f
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>