add efi partition to software based iscsi installs in efi mode
This expands the the iscsi root partition to be inline with
the other partition sizes of 256MB and allows for the creation
of an efi partition if booted in efi mode. Patch completes 99%
of the iscsi root installation on efi. The efi bootloader entry
in this scenario still has reliability issues and will be fixed
in a followup patch.

Change-Id: I7bb36a4a3b13f13a23a8e72f3e3028f2f5738cfd
Signed-off-by: Joey Boggs <jboggs@redhat.com>