installer: Don't reload, just re-parse
Previously ovirtnode.ovirtfunctions were used by the main thread (running
the UI) and the installer thread (which is a different thread
running in the backgrouund to keep the ui respsonsive). This lead to
problems with th eaug variable in ovirtfunctions. To work around this
problem, the installer thread is now only explicitly using
parse_defaults to re-parse the defaults file, instead of reloading the
whole ovirtfunctions module.

Change-Id: I368b6cf56e0ede48b6a564ed4cea274a9198a858
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>