service: Firstboot shouldn't block reboot
The sysd-upd-utmp-shutdown.service (see below)
is pulling in the var-log.mount unit.
That means it tries to mount /var/log as defined
in /etc/fstab (the entry was created as a part of
the installation). This fails because of some
reason (which is hard to debug, couldn't nail it
down).
But that doesn't matter because /var/log is
already mounted and working, just unknown to
systemd.

This patch unmounts /var/log so systemd can mount
it to prevent the reboot from failing.

rhbz#920208

Change-Id: I35779c9bb3dcf132d5555bdfa444e798b6b0f4cc
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>