ui: Add UIThreadConnection
This class can be used by other threads to interact/modify the UI
through the UI thread's mainloop.
If the interaction with the UI is not done through such a mechanism,
race conditions will (sooner or later) occur and can lead to all sorts
of UI errors (e.g. inconsitencies, incorrect rendering, ...).

This new class is directly used in the installer (thread).

rhbz#920049

Change-Id: I099a1fbb9106d5501c1f78c6274730b4801ccaf3
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>