net: Use ip instead of brctl
Previously brctl was used to gather informations about bridges, now the
ip command is used to do the same.
The intention is that we can drop bridge-utils once NetworkManager can
do the bridge stuff as well.

NotE: bridge-utils  is still kept because it's needed by the network
service to setup bridges.

rhbz#920125

Change-Id: I2022ce794ee2792e949457ad7664649c91702813
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>