make sure to only edit the Domain line in idmapd.conf
previously, if you set the domain to some value that is found
elsewhere in the idmapd.conf file, and then updated again, it would
replace all occurences of the value in the file.

This is especially bad if you do something like "nobody" since that
changes the user that nfsv4 runs as.

rhbz#962737

Change-Id: I6ceda931405e6c55152ac9044ff982c78095ad8b
Signed-off-by: Mike Burns <mburns@redhat.com>