synchronize log dir with iso on boot
During an upgrade, a service could be added that logs in a new
location under /var/log/<service>.  The directory will not exist
and start/stop of that service may fail.

This patch adds logic that will examing the contents of /var/log
on the base iso and create any directories found there but
not found in the HostVG-Logging partition.

Change-Id: Ife2399c9cb1fb7eb9fd572f9214ee74f8a61a8ce
Signed-off-by: Mike Burns <mburns@redhat.com>