app: Move plugins into their own namespace
Previously all plugins had to put the UI .py fies into the same
directory. This could lead to collisions and could became messy on the
long term. Now each plugin can create it's own folder in the base plugin
dirs for the setup and installer (their own namespace) to keep all
relevant files (wrt to the plugin) in one place and avoid collisions
with other plugins.

Now the discovery and initialization happens as follows:
1. Import all packages below a basepackage (e.g. ovirt.node.installer)
2. If the package has a function called 'createPlugins(application)'
then it is called. It is expected that this function creates all plugins
which shall be activated.

This way only the package is imported, the package/plugin itself is
responsible for all other imports.

Change-Id: Icaf10816d1c87d9d0404130f496f0f6cad13149a
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>