src: Restrict checks on new codebase
… and disable pylint for now.
Previously the whole codebase was checked, but this always failed,
because it's well known that the legacy codebase is not testable.
Therefor the checks are now only run on the new codebase. pylint has
been disabled for now, because it's to strict and we need to work out a
nice configuration which helps us.

Change-Id: I9650db5128306b24a45d5de59959f2ad1af97327
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>