fs: Introduce File and FakeFs
Two classes are added to allow unittesting without access to the
filesystem.
Previously unit tests were nearly impossible because of the many
informations which are retrieved from the filesystem (at runtime).
by wrapping the file i/o in a class a mocking framework (e.g.
python-mock) can be used to patch this file i/o class during unit tests
and redirect file i/o requests to a fake filesystem (e.g. an
in-memory fs).
FakeFs is a simple class implementing such a fake in-memory filesystem
which can be used during unit tests.

Change-Id: I7889240a956b359f2375955766f3cafcfbeaf0ee
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>
Signed-off-by: Mike Burns <mburns@redhat.com>