ui: Remove irritating text before opening dialogs
In the days of using less to display text on the screen we told the user
about pressing 'q' to quit. Now, in the days of using real dialogs, this
message (Press 'ESC' to quit) seems to be irritating, at first because
it's not shown within the dialog, but on e the page before, and second,
because there is already a close button on the dialog itself (in
addition to using ESC to close a dialog).

Change-Id: I8b264e02af8ea7ada10ac8f8d272763369b078bb
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>