network: Fix NIC.is_configured()
Previously a NIC was only assumed to be configured if a bootprotocol was
given. But this is wrong for the static IP cases, when no bootprotocol
but an IP address are set.
Thus the logic is updated and a NIC is assumed to be configured when a
bootprotocol or an IP address is set.

The symptoms were that the NIC appeared as unconfigured in the network
page, when a static IP was set.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=980465
Change-Id: Ic2d3e2d4ddb75958dd7cbb67bb29a305cc6d2411
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>