process: Make use of shell explicit
Previously some methods in the process module implicitly used the shell
keyword-argument, some others didn't.
If the shell=True arg is used, then Popen will allow and interprete a
string as the cmd, e.g. "ls /tmp" and will do the right thing. If
shell=False the Popen will look for a command named "ls /tmp". Because
in the case of shell=False, Popen expects a list, the first item in the
list beeing the command itself, and the subsequent items beeing
arguments to that command.

This patch removes the implicit use if shell=True to have the same usage
pattern accross all functions in the process module.
This prepares to solve problems e.g. related to kdump.

Summarized:
Args need to be a list in the default (shell=False) case:
process.call(["ls", "/tmp"])
This is a save way of calling a function with arguments, because quoting
the args is done by python.

Args need to be a string in the shell=True case:
process.call("ls /tmp | grep something", shell=True)
The string will be interpreted by a shell, thus all shell features can
be used in that string, but also quotation is up to the user and
therefor considered as unsafe.

Change-Id: I3cbc920e07fd4c3890028fe2f2d006e1be7c2ae6
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>