Remove $package-release file before creating it
When build the iso using autobuild.sh under ovirt-ndoe,
the package name will be ovirt-node-image. But in the
recipe/ovirt19-post.ks, the command:
ln -s /etc/system-release /etc/ovirt-node-image-release
creates link file ovirt-node-images-release that refers to
the /etc/system-release.

So this will cause both /etc/ovirt-ndoe-image-release and
/etc/system-release are link files and they refer to each other.

This patch remove the $package-release file anyway to avoid
refer to each.

Change-Id: I6da35c6a20de242d5c0c66ea06fc4edc07369b50
Signed-off-by: Wenyi Gao <wenyi@linux.vnet.ibm.com>