installer: Fix password page
Previously the installer accepted an empty password but failed to
complete the installation. This happened due to an incorrect password
check.
Now the password is checked correctly and the user get's feedback about
the different problems.
Additionally the installer only allows the user to carry on when a valid
password is given.

Change-Id: I2daeebd9ae5079d39af1b4f1b7be6eed9a9ce790
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=985275
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>