ui: Rework validation and exception handling
Previously the actual widget of a toolkit was responsible to call and
handle the validation stuff. The pros of this approach were that e.g.
handling simple validation errors was easy and could be directly
implemented by the widget. Problems arose with the introduction with
e.g. the ConfirmedEntry (which is a complex ui.Element in the UI
abstraktion layer). With the introduction of these complex ui.Elements
we also wanted - and needed - to pull the validation and exception
handling into the UI abstraction layer. This is necessary because now it
can be controlle din the UI abstraction layer how invalid data shall be
handled.

Change-Id: I4542a2641c6ce6f9f9126db608abca57a4ca453b
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>