installer: Fix EFI boot entry removal
Previously removing EFI entries during a reinstall failed due to an
incorrect call to efibootmgr.
This patch adds a wrapper around efibootmgr and provides an API to add
an remove EFI boot entries.

This shall prevent incorrect function calls to efibootmgr.
The new API is now also used in the legacy (installer) code.

Change-Id: I761ac056f4a168eefbc49319c0ff66239a9ded38
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=950553
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>