ui: No on_change on Table(multi) focus change
Previously an on_change event (which triggered an on_save() call) was
emitted when the focus in a Table with multiple-options changed from one
focus to the next. But this behaviour was incorrect.
NodePlugin.on_save() shall only be called when the data model change, so
when the user actualy changes data e.g. selects another entry or
deselects one.
The on_change is now only called - in a user with multi=True - when the
user selects or deselects an entry.
To capture focus shifts/changing entries the Table.on_change signals has
to be used directly.

Change-Id: Ic6d9fc2ca9dbcbb4ee009ebc3b7e342a9e44c66c
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>