network: Fix Bonding.delete_all()
Previously the class tried to read a file which is only available when
the bonding module is loaded. This could lead to an exception when there
was a "greedy" call to Bonding.delete_all() (so not checking if bonding
was enabled at all).
This is onw fixed by only reading the file when bonding is enabled.

Change-Id: Iba592a6e02fd036283fec899e9081edabc941e2b
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>