ssh: Show consistent transaction
Previously the transaction elements in appearing in the  transaction
used on the security page weren't obviously connected to the changes
done in the TUI.
Now the same transaction elements are used in the transaction.

Sometimes it's not possible to distinguish (in the defaults class) what
value changed. Therefor we are running all transaction elements and
don't try to be smart.
It could be solved by handling this in the TUI layer.

Change-Id: Idff93c3d8a59814941d1ed34d92c78c97a39d70d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=982425
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>