Validate "Other Device" against live device
Previously, we didn't check whether or not "Other Device" input
matched against a known live device or not (USB boot, etc). Do so.
As part of this, slight rearrangement of storage.Device to improve
responsiveness when looking for udev devices (we're assuming udev
will not change during the install), and get the new validation
workflow to properly invalid pages if a plugin's on_change throws
an exception.

Change-Id: I44b87876876965418765cbb6a0c0030bf143f156
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=885017
Signed-off-by: Ryan Barry <rbarry@redhat.com>