boot: Add savedefault to primary boot entry
Previously savedefault was only set on the backup (second) boot entry of
the grub cfg. That lead to problems when you wanted to set the primary
entry the default after booting the second one:
1. Boot primary entry (is default)
2. Boot second entry (becomes default, as of savedefault)
3. Reboot - Boots into second entry
4. Boot primary entry (does _not_ become default)
5. Reboot - Boots into second entry
This patch also adds savedefault to the primary boot entry, that way the
primary entry will be set as default when it is selected in the boot
entry. That would lead to the following behavior:
...
4. Boot primary entry (becomes default, as of new savedefault)
5. Reboot - Boots into primary entry

Change-Id: I49fcc97af7ac275086dd10ac26f9edad90e7aea9
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1017114
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>