Manually set a GID for CIM
Previously, we added a new group for CIM in %post before creating
the admin user, which creates ownership problems on upgrades.
Instead, see whether or not the cim user exists, and try to take
gid 501. If we can't get 501, create with any GID (plugin being
installed after the fact? Planning for live plugins and users
added).

Change-Id: If5d235093a84c2a8b9a3dfc220e50565b7c7dc62
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1018063
Signed-off-by: Ryan Barry <rbarry@redhat.com>