init: Fix link detection for auto-installation
The link state of a device needs to be detected when BOOTIF=link is
given on an auto-install.
Previously the link state of a device was determined in one pass. This
could lead to problems when it takes the hardware a while to initialize
or at least set the link state to UP.
Now five (5) passes are allowed to find a device with a link.

Change-Id: I72307c6ea53002014c680087166e8a64745dd3fd
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=959957
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>