Fix devpath and lspci parsing
Previously the assumptions about a devpath were wrong - in some
situations, the device addr is at the end of the path, because it might be
leaded by the path of a PCI bridge.
Now the device PCI addr are taken from the back of the devpath, this is
more robust. Furthermore the parsing of the lspci output got improved.

Change-Id: I59678c97fe5192641df7616ade03afbdd4c745e6
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1046523
Signed-off-by: Fabian Deutsch <fabiand@redhat.com>