migration: Move migration script to boot
We reduce the config migration to one point - during boot.
The script is run once, this is ensure by only running this script if
OVIRT_CONFIG_VERSION is not present in /etc/default/ovirt.
In future we need to enhance this logic.

The previous calls to the migration script were removed to simplify the
testing.

Change-Id: Id3dce11ce0005ddc12d9fd2fec2792ff1e87513c
Signed-off-by: Fabian Deutsch <fabiand@redhat.com>
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>