network: Remove biosdevname conversion code
In the beginning of biosdevname, affected NICs names had to be
converted from their "old" name to the "new" biosdevname name.

But as biosdevname is out of the door for a long itme, we can drop that
code path.

A note on RHEV-H: Only upgrades from the previously released version are
supported, that menas we don't need to support an upgrade from 6.3 to
6.5.

Change-Id: I780dd0770d9f43be6ca0882341bb9fa97d179f96
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1065256
Signed-off-by: Fabian Deutsch <fabiand@redhat.com>