Add ovirt-node-config
ovirt-node-config is a commandline tool to configure Node.
Basically the tool exposes the classes of ovirt-node.config.defaults.
That way the system can basically configured using the commandline. The
same codepaths are shared between the CLI and TUI.

This is mainly intended for quality issue to check all codepaths.

Some examples:

General help:

$ ovirt-node-config -h

Help on a class:

$ ovirt-node-config help IPv6

Help on a function:

$ ovirt-node-config help IPv6.configure_dhcp

Run a function:

$ ovirt-node-config run IPv6.configure_dhcp

$ ovirt-node-config run Network.configure_static ens3 1.2.3.4 255.0.0.0
1.2.3.254

Change-Id: I0d06244e1380b4880381757074ad19c5d96a7771
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>