setup: Disable network page when managed
Previously the suer could edit the network connections even when the
system was managed.
This could lead to problems when the management instance did a network
configuration which wasn'tunderstood by the TUI.
To prevent these kind of problems the NIC details dialog is now disabled
when any NIC on the Node is managed.

Change-Id: Ia0ec61ec3c655fea0b8274519e105e4c7a74409f
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1065385
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>