Make the logrotate interval configurable
Previously, we only gave the option for size to be checked. Let
the period be checked as well, users can rotate logs daily if
desired. "size" and intervals are exclusive in EL6 other than
"minsize", so we can't get to the desired conclusion, but we can
prep by adding an option in the UI.

Change-Id: Ifb2639225f8b793b302e36a2c7639cb0769d4df4
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1046417
Signed-off-by: Ryan Barry <rbarry@redhat.com>