Apply default disable on network details page
We should check for the value being "none" in case of users
manually selection the "Disabled" checkbox as well as None in case
they never picked it at all. This way, we can avoid a situation
where users leave auto/dhcp enabled by failing to make a selection
for that field and saving the page.

Change-Id: I406f250e7cd2f3718d063826dcbe48cb1cbd57ef
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1094552
Signed-off-by: Ryan Barry <rbarry@redhat.com>