mount_config: fix non-default directory persistence
Currently mount_config fails to restore directories that are not
present in the original image of the root filesystem. E.g.

    mkdir -p /var/lib/vdsm/foobar/moo
    echo "123" > /var/lib/vdsm/foobar/moo/cow.txt
    persist /var/lib/vdsm/foobar

Then, after reboot, ovirt_start_early will call mount_config which
will result in
    mkdir -p /var/lib/vdsm
    touch foobar
    mount -n --bind /config/var/lib/vdsm/foobar /var/lib/vdsm/foobar

This will fail because one can't bind mount a directory on a file.

This patch fixes this issue.

Change-Id: Ifa2b854980ee28cce5d55e0faf30442e6c01a4b8
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>