mount_config: Add support for symbolic links
Up until now, symbolic links used to be treated like regular files
(just as directories were before the preceding patch of the branch).
This meant that trying to restore a persisted symlink would try to
bind mount whatever the symbolic link was pointing at originally.
However, some application might rely on a symbolic link being a
symbolic link and not whatever that symbolic link was pointing at.

With this patch, if we persisted a symbolic link as:

    /config/files:
        /absolute/path/to/the/symbolic/link

    /config/absolute/path/to/the/symbolic/link linking to:
        /absolute/path/to/target/of/symbolic/link

we'll get

    /absolute/path/to/the/symbolic/link linking to:
        /absolute/path/to/target/of/symbolic/link

Once the persist is fixed for symbolic links, it will behave like
file persistence does now, you should call persist after changes
to the link.

Change-Id: I8e009e3039b2df80d57e60460d6cc888a91f7577
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>