ui: Improve the setting of the console path
Previously the exampel did not match what was required to be put into
the field (previously only the device [ttyS0] was required, but the
example stated a full path [/dev/ttyS0]).
This is now corrected, furthermore was the detection of the correct grub
cfg file reordered to be quicker in many cases (the discovery by fs
label was pulled to the end).

Change-Id: I8f88e8c0f5a73400728440339dd6619982c5aea8
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>