edit-node: Identify packages with * in --install
The --install option accepts packages that will be updated
in the oVirt Node iso. We can accept packages in the below
format:

./edit-node --repo=./myrepo.repp
            --install vdsm,ovirt-node-plugin-vdsm* base_iso.iso
            --name mynewiso

However, edit-node is not able to generate the plugins
files correctly in /etc/ovirt-plugins.d/<plugin-name>
when the * is used affecting the Plugins option in the TUI.

This patch adds the hability to edit-node to identify
the packages correctly when * is used for similar packages name.

Change-Id: Iee564235caacffe04bafde0a3fb29404d135c2a2
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1119606
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>