Rework the inputs on the RHN page a little
Don't blindly set the text on input fields. Instead, we should
disable input fields and keep their value, passing it back in
if it's re-enabled. Additionally, this prevents getting into a
situation where bad input in rhn.url or rhn.ca can be marked as
valid by changing the type from Satellite to SAM, as well as
improperly marking the page as invalid if invalid input has been
entered into fields which are no longer active.

Change-Id: I977ac638914db789bc367a91cdbfda3851d2e4eb
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1064836
Signed-off-by: Ryan Barry <rbarry@redhat.com>