Make EL7 register to SAM instead of RHN
rhnreg is a no-go on EL7. Use subscription-manager instead. This
is a workaround until the RHN code can be redone and modularized.
Additionally, properly set RHN_TYPE if no selection has been
made and there's nothing pending in the model.

This does not resolve the bug since a product certificate is
needed for subscription-manager to return 0, but allows
registration

Change-Id: I73c05c154a21a68b1fd99889a05b3f5abd7d98d9
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1120558
Signed-off-by: Ryan Barry <rbarry@redhat.com>