Fix persistence logic
Previously the return code of a list of files which needed to be
persisted, was determined by the last file to be persisted. And even if
there was no severe issue, it could lead to a situation where the retun
code was unequal 0, which returned to a failed auto-installation.

This patch also introduces some better descriptions for failed
auto-installations.

Change-Id: I7749fb2a423283c5748cbe990374d4790cffeb99
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1120650
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>