firewall: Use correct logic if firewalld is used
Previously the decision if firewealld should be used or not was lax, it
was only based on the fact that firewalld was used.
Now the decision is made on the install state and service state of
firewalld.

Now firewalld will only be configured if it's installed and the service
running, otherwise iptables will be used.

Change-Id: Ica13f63ac67d0ba1dd29f5b1566e96036e78eefc
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1150328
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>