Make rhn_model work with Satellite 6
First, don't try --autosubscribe, since it's caused problems with
SAM and Satellite 6. Don't remove it, but leave it disabled until
we clarify what's supposed to happen here.

If an organization is specified, use it. Check whether the call
prompts for one to see if we're trying to hit Sat6 without
specificying one.

Use process.pipe instead of check_output so we don't wait forever
if it is prompting.

Don't use wget. It shifts the TUI. urllib.urlrequest is an easy to
use pythonic replacement.

Change-Id: Ifc5f87400efa351c8fa4cd4ced9fc8b4462e9d2e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1067355
Signed-off-by: Ryan Barry <rbarry@redhat.com>