recipe: Add sane multipath conf
Previously we used an empyt multipath.conf.
This lead to a situation where devices were claimed by multipath, even
if they weren't multipath devices.
To fix this, we now provide a configuration file which set's the
relevant directive, to only mark verified multipath devices as such.
This basically means that mpath devices will only be created for devices
with multiple paths.

Change-Id: I7bf827830450acd0aba60cedba259a2a44ac23d6
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1152948
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>