spec/recipe: Only update mpath.conf once
Previously we updated multipath.conf in the spec and in the recipe, this
lead to a situation where we created an invalid conf.
Which implied that multipath did not work at all.
Now the file is only modified in the recipe.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1152948
Change-Id: I3fdb8d0061310144c09280a374d8bc0c1a53b3cd
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>
(cherry picked from commit dcde4b6bc95c96d6c90897f9b45646a2770a2f45)