installer: Be more conservative with partprobe
Previously partprobe was called in many places to update the partition
device links.
The initial cause for the missing device links were problems with udev
and multipath.
Now that these bugs have been addressed, the partprobe usage is known to
be buggy when layered dm devices exist or something is mounted. Thus we
remove the partprobe usage to reduce the probability for problems caused
by an excessive use.

Change-Id: I713e83009e6c623078ef30ffdbd0a620515cf8de
BNug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1167240
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>