recipe: More robust multipath configuration
Previously some multipath configuration snippet got appended to the
existing configuration file.
This could lead to a situation where the defaults section appeared
twice, this made the configuration invalid, and multipath would not work
at all.

Now a different approach is used to ensure that a vlid configuration is
written.

Change-Id: I1e878b8cd2ce4b456380300dca2fa51c92523547
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>