installer: Name multipathed device for boot
Previously (pre-el7) it was tried to discover and assemble the correct
multipath devices in dracut - this worked quite well.
In el7 we need to name multipath devices we want to get assembled in
dracut, if not, then the assmebling can fail due to races with dracut
and udev.
To prevent this, now the wwid of the multipathed root device is given on
the kernel cmdline, this is then pickedup within dracut to assemble the
correct device.

Change-Id: I13b9cd1c07b0780813ba91cf75a7ed0070c306d0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1152948
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>