installer: Use different size source
Previously sfdisk was used to determin the size of disks. This could
fail if the disk contained a GPT label (sfdisk does not support GPT),
this does actually not matter (what label is used) because we only need
the size of the disk, but we still used sfdisk.
Now lsblk is used, which only looks at the disk itself, and does not
take the label into account.

Change-Id: I913ee34a50f9a79e51d4a8b073318c4966576dd3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1174146
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>