Bump crashkernel size to 256M on EL7
To prep for dmsquash-live changes for kdump, bump the crashkernel
size up. Parameterize that in common-install, and convert
common-efi to a parameterized file so the installer options match
on EFI and non-EFI systems.

Also, don't use rd.lvm=0 on EL7 -- it prevents the kdump kernel
from mounting the local dump target.

Change-Id: I4e8f051bf4aa5cec138143d51048de22dcd68105
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1180371
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1180377
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1095140
Signed-off-by: Ryan Barry <rbarry@redhat.com>