ovirt-init-functions: Grab 06 or 07 bytes for BOOTIF karg
BOOTIF kernel argument takes the network interface as parameter
or the 06 bytes of mac address using dashs between the bytes.
Currently, the ovirt-init-functions.sh expects 07 bytes instead of 06 bytes of mac
address because PXE IPAPPEND always appends "01-" to the front of the MAC.
However, if user provides the 06 bytes of MAC address manually to the GRUB to
append to the boot the autoinstall doesn't work.
This patch adds the "or" statement for validate 06 or 07 bytes during the
ovirt-node boot.

Change-Id: I8bfcc8ddabc74ee67d1a175fd4824ed9173c0215
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>