ntpd: condition for start
Due the limitation that all services in the oVirt Node
must be enabled during the build time and there is
no option to disable the pre-configured ntp servers
from ntp.conf this patch adds a condition for
start of ntp service. The condition is the
/etc/ntp.conf be persisted. In other words, if users
set NTP servers via TUI or auto-install using
ntp=my-preferred-ntp-server will enable the ntp
daemon to be started in the boot.

Change-Id: Iebee8a8581fd0b51d07bde04b1f878a1cb5481ba
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1061081
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>