Always restart the kdump service, even if kdumpctl is present
Restart kdump no matter what, so the output of "systemctl status
kdump.service" or "service kdump status" matches what kdumpctl
already knows, and the output is captured in the journal.

Leave "kdumpctl restart" anyway, since "service kdump restart"
does not pass failures back, and we want to catch the exception
so we know, even if this means duplicating a little work.

Change-Id: I187a9b3642dc1ee4b353b3adfdb051c58545fe01
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1180371
Signed-off-by: Ryan Barry <rbarry@redhat.com>