Use "ssh" instead of "net" in the augeas template on EL7
Since augeas has support now, and using "ssh" in kdump.conf is
required to make ssh kdumps work on EL7 (but we can't write from
a template without breaking engine features), use "net" and sub
out the key to "ssh" if it's EL7.

This is a temporary fix until augeas support is backported to EL6,
when we can just use "ssh" everywhere.

Change-Id: I0c2b73726354cfaea477d8e041a86914a6dd89fe
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1180371
Signed-off-by: Ryan Barry <rbarry@redhat.com>