defaults: Add vdsmd to stop/start in network conf
When configuring network the listed services are stopped
and started: network, ntpd, ntpdate, rpcbind, nfslock,
rpcidmapd, nfs-idmapd and rpcgssd. VDSM requires rpcbind
and if we stop this service during the configure of
network vdsm will be terminated as it requires this
service in systemd unit file. This patch adds the
vdsmd as a service to be stopped and started when
network is configured.

Change-Id: I11808417c8358777478f0cf9f821d93cafbb9f3b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1209836
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>